Exploring dicom server functionality in scripts
Using pydicom and pynetdicom to create a dicom server, the functionality
of which is explained very well in the following link:
https://book.orthanc-server.com/dicom-guide.html#dicom-guide

Such a dicom server can be recognised and interacted with natively by
the dicom viewers used by radiologist (eg.  Horus or OsiriX). This alows
users to send their scans to our server without leaving the environment
they are comfortable in. Alternatively the deep_dolphin server could
routinely ping any hospital database and process new scans. Both of
these methods remove points of friction for the end users.