Started refactoring ContourBuilder
Must be able to recognise multiple contours within one slice. This
functionality will be included in the SliceCountourBuilder class.

While testing in mask_to_dicom_exporter_test.py I discovered a problem
with the algorithm currently utilised by ContourGenerator class.

Need to implement the ability to backtrack if a wrong path is followed.