ContourBuilder functionality fully implemented
Introduced the ability for the ContourBuilder to backtrack on exploration
if it encounters a 'dead end'. This significantly increases reliability
and allows the ContourBuilder to be looped until a complete set of points
have been explored. This functionality will soon be implemented in the
SliceContourBuilder and is now demonstrated in mask_to_dicom_exporter_test.py

Code still needs signifficant cleaning but functionality is there.