Fully implemented conversion of mask slice to contours
SliceToContourConverter class accepts a mask slice and uses the EdgeDetector
to create a set of edge points. This set is unordered and therefore must be
fed into the PointsToContourConverter to generate a set of contours.

PointsToContourConverter applies the algorithm discussed in the following link to
find contours that contain all of the inputed points (for the most
part). Detailed documentation coming soon.

https://towardsdatascience.com/the-concave-hull-c649795c0f0f