Begun implementation of SliceToContourConverter tests
For each slice of a nii mask we check that the area described by the
contours is within a certain error margin of the slice area.

Already found and attempted to fix certain bugs through this test.

One of these fixes was the inclusion of a maximimum_side_length parameter
for the SetToContourConverter class. As the name suggests, this parameter
sets a maximum distance between vertices. This distance can be set depending
on the context.

For our context where we hve very tightly packed points a distance of 4
is appropriate and has therefore been set as a default.